Preserve old behavior.
